Move CompWrapper from arrays/base.py to ops.py
