+++ pandas/core/arrays/__init__.py
@@ -1,8 +1,7 @@
 from .array_ import array  # noqa
 from .base import (ExtensionArray,    # noqa
                    ExtensionOpsMixin,
-                   ExtensionScalarOpsMixin,
-                   CompWrapper)
+                   ExtensionScalarOpsMixin)
 from .categorical import Categorical  # noqa
 from .datetimes import DatetimeArray  # noqa
 from .interval import IntervalArray  # noqa
